Fix array_merge fatal error in Magento 2.3.1
(Magento 2.3.1, PHP 7.2, default mode, PHP errors not hidden). On a site which is not using custom currency, any time a price is accessed, Model\Config::_getCurrency() returns false, and so array_merge fails with 'Argument #1 is not an array'. All the calls to this allow for a false return value, and this fix appears to work for custom and non custom currency.